requirements.txt
selenium
newspaper3k
bs4
nytimes-scraper